// FE/super-app/modules/shared/src/lib/interfaces/i-form-model.ts
export type FormValidation = 'required' | 'email' | 'min' | 'max' | 'pattern' | 'custom';
export type FieldType = 'textbox' | 'number' | 'dropdown' | 'date' | 'checkbox' | 'radio' | 'textarea' | 'editor' | 'address' | 'percent' | 'chips' | 'autocomplete' | 'radiobuttonlist' | 'checkboxlist' | 'table' | 'container';
export type LayoutWidth = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12;
export type ControlFilter = Filter[] | Func<Filter[]> | Promise<Filter[]>;
export type PrimitiveType = string | Date | number;

// Interface cơ bản cho validation
export interface IFormValidation {
    type: FormValidation;
    message: string;
    value?: any;
}

// Interface cơ bản cho field
export interface IFormField {
    /** Tên field, dùng để định danh field trong form */
    field: string;
    /** Label hiển thị cho field */
    label?: string;
    /** Tooltip hiển thị cho field */
    fullLabel?: string;
    /** Kiểu dữ liệu của fiel    /** Giá trị của field */
    value?: any;
    /** Danh sách các validation rule */
    validations?: IFormValidation[];
    /** Field có bị disable hay không */
    disabled?: boolean;
    /** Kiểu dữ liệu của field */
    dataType: DataType;
    /** Có localize không */
    useLocalization: boolean;

    //#region Form
    hiddenInForm: boolean;
    /** Placeholder text khi field trống */
    placeholder?: string;
    /** Số cột chiếm trên màn hình web (1-12) hoặc độ rộng cố định (px) */
    layoutWidth: LayoutWidth | string;
    /** Số cột chiếm trên màn hình điện thoại (1-12) hoặc độ rộng cố định (px) */
    mobileWidth: LayoutWidth | string;
    //#endregion

    //#region List
    cellTemplate?: TemplateRef<any> | undefined;
    hiddenInList: boolean;
    includeInList: boolean;
    /** Số cột chiếm trên giao diện list (1-12) hoặc độ rộng cố định (px) */
    columnWidth?: number | string;
    allowFilter: boolean;
    allowSort: boolean;
    /** Placeholder text khi field trống */
    searchPlaceholder?: string;
    filterOperator?: FilterOperator;
    headerClass?: string;
    bodyClass?: string;
    separator: string; // Trong trường hợp field là reference
    pipe?: PipeTransform;
    pipeParams?: any[];
    asyncPipe: boolean;
    //#endregion

    /** Tạo bản sao của field */
    clone(): this;
}